+++ authentication/src/main/kotlin/com/uber/sdk2/auth/api/request/CrossApp.kt
@@ -1,0 +1,28 @@
+/*
+ * Copyright (C) 2024. Uber Technologies
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.uber.sdk2.auth.api.request
+
+/** Provides different apps that could be used for authentication using SSO flow. */
+sealed class CrossApp {
+  /** The Eats app. */
+  object Eats : CrossApp()
+
+  /** The Rider app. */
+  object Rider : CrossApp()
+
+  /** The Driver app. */
+  object Driver : CrossApp()
+}
